Move server to server directory
