+++ server/tests/routes_test.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from main import app, talisman
+from server import app, talisman
 import pytest
